hotfix: renamed folder and just changed the cli cmd
